dom/html/test/test_bug519987.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=519987
-->
<head>
  <title>Test for Bug 519987</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=519987">Mozilla Bug 519987</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 519987 **/
var xmlns = 'http://www.w3.org/1999/xhtml';
is((new Image()).namespaceURI, xmlns, "Unexpected namespace for new Image()");
is((new Audio()).namespaceURI, xmlns, "Unexpected namespace for new Audio()");
var titles = document.getElementsByTagName("title");
var t = titles[0];
t.parentNode.removeChild(t);
document.title = "abcdefg";
is(titles[0].namespaceURI, xmlns, "Unexpected namespace for new <title>");

</script>
</pre>
</body>
</html>
